Done2

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/RoleDao.java
@@ -6,5 +6,5 @@ import ru.kata.spring.boot_security.demo.model.Role;
 import ru.kata.spring.boot_security.demo.model.User;
 
 @Repository
-public interface RoleDao extends JpaRepository<Role, Integer> { }
+public interface RoleDao extends JpaRepository<Role, Long> { }
